installing the package in one command on all Supervisors.
Install RPC installs on both Supervisors and returns cluster_validated in Validated
message as true on success. It returns NOT_SUPPORTED_ON_BACKUP in InstallError
when asked to install on standby supervisor only. Image is present after
Install on both Supervisors or neither Supervisor. Such systems do not
need to sync the package from the peer Supervisor.

Activate RPC activates on both Supervisors and returns cluster_activated as true
in ActivateOK message on success. It returns NOT_SUPPORTED_ON_BACKUP in
ActivateError message when asked to activate on standby supervisors only.
Such systems run the package by rebooting the entire system and do not
need to switchover to the peer Supervisor before rebooting a Supervisor.

Added change to request snapshot of image and config to support auto recovery
for systems which support snapshot to secondary disk on each Supervisor.
When post upgrade reboot early init failures cause the system to boot from the
second disk, the system auto recovers to the previously running image and config.
This is specified by InstallRequest snapshot_to_secondary_disk.