system: add icmp_ext_data field to TracerouteResponse
In order to get the MPLS key/value pairs, traceroute must run
with the '-e' flag.
However, this flag makes traceroute include all ICMP extensions,
and only the MPLS is available in a parsed format.
For other ICMP extensions, they will be returned in a format
"CLASS/TYPE: followed by a hexadecimal dump" as per the
traceroute man page.
This change introduces a way to return stream this data back to the
client.